+++ app/Core/Cache/CacheInterface.php
@@ -1,0 +1,24 @@
+<?php declare(strict_types=1);
+/*
+ Rqwatch
+ Copyright (C) 2026 Giannis Kapetanakis
+
+ This Source Code Form is subject to the terms of the Mozilla Public
+ License, v. 2.0. If a copy of the MPL was not distributed with this
+ file, You can obtain one at http://mozilla.org/MPL/2.0/.
+*/
+
+namespace App\Core\Cache;
+
+interface CacheInterface
+{
+	/*
+	public function get(string $key): mixed;
+
+	public function set(string $key, mixed $value, ?int $ttl = null): void;
+
+	public function delete(string $key): void;
+
+	public function has(string $key): bool;
+	*/
+}