+++ CHANGELOG.md
@@ -3,6 +3,7 @@
 ## Master Dev Branch
 
 ### 2026-08-07
+- Add CacheInterface and RedisCache
 - Add responsive hamburger menu and fixed menu widths
 
 ## Version v1.8.3 - Released: 2026-08-06